refactor(protocol): move PublishStreamConnectionResponse classes to own folder
