+++ packages/protocol/src/protocol/control_layer/index.ts
@@ -3,7 +3,7 @@ import ErrorResponse from "./error_response/ErrorResponse"
 import { ErrorCode } from "./error_response/ErrorResponse"
 import PublishRequest from "./publish_request/PublishRequest"
 import PublishStreamConnectionRequest from "./publish_stream_connection_request/PublishStreamConnectionRequest"
-import PublishStreamConnectionResponse from "./publish_stream_connection_request/PublishStreamConnectionResponse"
+import PublishStreamConnectionResponse from "./publish_stream_connection_response/PublishStreamConnectionResponse"
 import ResendFromRequest from "./resend_request/ResendFromRequest"
 import ResendLastRequest from "./resend_request/ResendLastRequest"
 import ResendRangeRequest from "./resend_request/ResendRangeRequest"
@@ -49,8 +49,8 @@ import './unsubscribe_response/UnsubscribeResponseSerializerV1'
 import './unsubscribe_response/UnsubscribeResponseSerializerV2'
 import './publish_stream_connection_request/PublishStreamConnectionRequestSerializerV1'
 import './publish_stream_connection_request/PublishStreamConnectionRequestSerializerV2'
-import './publish_stream_connection_request/PublishStreamConnectionResponseSerializerV1'
-import './publish_stream_connection_request/PublishStreamConnectionResponseSerializerV2'
+import './publish_stream_connection_response/PublishStreamConnectionResponseSerializerV1'
+import './publish_stream_connection_response/PublishStreamConnectionResponseSerializerV2'
 
 export {
     BroadcastMessage,
