+++ packages/protocol/src/protocol/control_layer/publish_stream_connection_response/PublishStreamConnectionResponseSerializerV2.ts
